+++ swoole_websocket_server.c
@@ -240,7 +240,7 @@ void swoole_websocket_onOpen(http_context *ctx)
 
         if (SwooleG.enable_coroutine)
         {
-            int ret = sw_coro_create(fci_cache, 2, args, retval);
+            long ret = sw_coro_create(fci_cache, 2, args, retval);
             if (ret < 0)
             {
                 SwooleG.serv->factory.end(&SwooleG.serv->factory, fd);
@@ -389,7 +389,7 @@ int swoole_websocket_onMessage(swEventData *req)
 
     if (SwooleG.enable_coroutine)
     {
-        int ret = sw_coro_create(fci_cache, 2, args, retval);
+        long ret = sw_coro_create(fci_cache, 2, args, retval);
         if (ret < 0)
         {
             zval_ptr_dtor(zdata);
